+++ SRTPluginProviderRE5/PluginInfo.cs
