--- SRTPluginProviderRE5/SRTPluginProviderRE5.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SRTPluginProviderRE5", "SRTPluginProviderRE5\SRTPluginProviderRE5.csproj", "{FB803174-8A74-4CC0-8645-B1D6C4F53288}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SRTPluginBase", "..\SRTHost\SRTPluginBase\SRTPluginBase.csproj", "{431972BA-04CE-4D4B-8ECE-80D351B02868}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProcessMemory", "..\ProcessMemory\ProcessMemory\ProcessMemory.csproj", "{C43019D3-C173-4C82-B8A0-6927CE5FCA38}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FB803174-8A74-4CC0-8645-B1D6C4F53288}.Debug|x86.ActiveCfg = Debug|x86
-		{FB803174-8A74-4CC0-8645-B1D6C4F53288}.Debug|x86.Build.0 = Debug|x86
-		{FB803174-8A74-4CC0-8645-B1D6C4F53288}.Release|x86.ActiveCfg = Release|x86
-		{FB803174-8A74-4CC0-8645-B1D6C4F53288}.Release|x86.Build.0 = Release|x86
-		{431972BA-04CE-4D4B-8ECE-80D351B02868}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{431972BA-04CE-4D4B-8ECE-80D351B02868}.Debug|x86.Build.0 = Debug|Any CPU
-		{431972BA-04CE-4D4B-8ECE-80D351B02868}.Release|x86.ActiveCfg = Release|Any CPU
-		{431972BA-04CE-4D4B-8ECE-80D351B02868}.Release|x86.Build.0 = Release|Any CPU
-		{C43019D3-C173-4C82-B8A0-6927CE5FCA38}.Debug|x86.ActiveCfg = Debug|x86
-		{C43019D3-C173-4C82-B8A0-6927CE5FCA38}.Debug|x86.Build.0 = Debug|x86
-		{C43019D3-C173-4C82-B8A0-6927CE5FCA38}.Release|x86.ActiveCfg = Release|x86
-		{C43019D3-C173-4C82-B8A0-6927CE5FCA38}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9F1CA3A3-4EED-43A9-B230-BC60B2E2ECE9}
-	EndGlobalSection
-EndGlobal
