+++ SRTPluginProviderRE5/Structs/GameStructs/GamePlayerHP.cs
